# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: |-
    WASAText messaging platform API.
    Connect with your friends effortlessly using WASAText! Send and receive messages,
    whether one-on-one or in groups, all from the convenience of your PC.
  version: "1.0.0"
  contact:
    name: WASAText Team
    email: support@wasatext.com
...
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /groups/{groupId}/members:
    parameters:
      - name: groupId
        in: path
        required: true
        description: Unique identifier of the group
        schema:
          type: string
          pattern: '^[a-zA-Z0-9_-]+$'
          minLength: 1
          maxLength: 255
        example: "group123456"
    post:
      tags: ["groups"]
      summary: Add user to group
      description: Adds a user to an existing group
      operationId: addToGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Add user to group request
              properties:
                userId:
                  type: string
                  description: ID of the user to add to the group
                  example: "user789012"
                  pattern: '^[a-zA-Z0-9_-]+$'
                  minLength: 1
                  maxLength: 255
              required:
                - userId
      responses:
        '201':
          description: User added to group successfully
          content:
            application/json:
              schema:
                type: object
                description: Add member confirmation
                properties:
                  success:
                    type: boolean
                    description: Indicates if the operation was successful
                    example: true
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Group or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Use the user identifier returned from login
  schemas:
    Error:
      type: object
      description: Error response object
      properties:
        message:
          type: string
          description: Human-readable error message
          example: "Error message"
          minLength: 1
          maxLength: 1000
        code:
          type: string
          description: Machine-readable error code
          example: "ERROR_CODE"
          pattern: '^[A-Z_]+$'
          minLength: 1
          maxLength: 50
      required:
        - message
...